Introduced TestRunRecorder (which is not used yet).
