+++ src/main/java/net/jqwik/discovery/PropertyMethodResolver.java
@@ -3,6 +3,7 @@ package net.jqwik.discovery;
 import java.lang.reflect.*;
 import java.util.*;
 
+import net.jqwik.recording.*;
 import org.junit.platform.commons.support.*;
 import org.junit.platform.engine.*;
