+++ src/main/java/net/jqwik/execution/JqwikExecutor.java
@@ -1,24 +1,28 @@
 package net.jqwik.execution;
 
+import java.util.function.*;
+import java.util.logging.*;
+
+import org.junit.platform.engine.*;
+
 import net.jqwik.api.*;
 import net.jqwik.descriptor.*;
 import net.jqwik.execution.pipeline.*;
-import org.junit.platform.engine.*;
-
-import java.util.function.*;
-import java.util.logging.*;
+import net.jqwik.recording.*;
 
 public class JqwikExecutor {
 
 	private final LifecycleRegistry registry;
+	private final TestRunRecorder recorder;
 	private final PropertyTaskCreator propertyTaskCreator = new PropertyTaskCreator();
 	private final ContainerTaskCreator containerTaskCreator = new ContainerTaskCreator();
 	private final ExecutionTaskCreator childTaskCreator = this::createTask;
 
 	private static final Logger LOG = Logger.getLogger(JqwikExecutor.class.getName());
 
-	public JqwikExecutor(LifecycleRegistry registry) {
+	public JqwikExecutor(LifecycleRegistry registry, TestRunRecorder recorder) {
 		this.registry = registry;
+		this.recorder = recorder;
 	}
 
 	public void execute(ExecutionRequest request, TestDescriptor descriptor) {
@@ -41,17 +45,14 @@ public class JqwikExecutor {
 		if (descriptor.getClass().equals(SkipExecutionDecorator.class)) {
 			return createSkippingTask((SkipExecutionDecorator) descriptor, pipeline);
 		}
-		return ExecutionTask.from(
-			listener -> LOG.warning(() -> String.format("Cannot execute descriptor [%s]", descriptor)),
-			descriptor.getUniqueId(), "log warning");
+		return ExecutionTask.from(listener -> LOG.warning(() -> String.format("Cannot execute descriptor [%s]", descriptor)),
+				descriptor.getUniqueId(), "log warning");
 	}
 
 	private ExecutionTask createSkippingTask(SkipExecutionDecorator descriptor, Pipeline pipeline) {
 		String taskDescription = String.format("Skipping [%s] due to: %s", descriptor.getDisplayName(), descriptor.getSkippingReason());
-		return ExecutionTask.from(
-			listener -> listener.executionSkipped(descriptor, descriptor.getSkippingReason()),
-			descriptor.getUniqueId(),
-			taskDescription);
+		return ExecutionTask.from(listener -> listener.executionSkipped(descriptor, descriptor.getSkippingReason()),
+				descriptor.getUniqueId(), taskDescription);
 	}
 
 	private ExecutionTask createPropertyTask(PropertyMethodDescriptor propertyMethodDescriptor, Pipeline pipeline) {
