+++ src/main/java/net/jqwik/recording/TestRunRecorder.java
@@ -1,0 +1,12 @@
+package net.jqwik.recording;
+
+public interface TestRunRecorder {
+	void record(TestRun testRun);
+
+	TestRunRecorder NULL = new TestRunRecorder() {
+		@Override
+		public void record(TestRun testRun) {
+
+		}
+	};
+}
